tests/test_main.py
```python
import unittest
from unittest.mock import patch
from src import main

class TestMain(unittest.TestCase):

    @patch('src.main.integrateChatbot')
    def test_integrateChatbot(self, mock_integrateChatbot):
        main.integrateChatbot()
        mock_integrateChatbot.assert_called_once()

    @patch('src.main.generateResponse')
    def test_generateResponse(self, mock_generateResponse):
        main.generateResponse("Hello")
        mock_generateResponse.assert_called_once_with("Hello")

    @patch('src.main.customizeResponse')
    def test_customizeResponse(self, mock_customizeResponse):
        main.customizeResponse("Hello", {"tone": "friendly"})
        mock_customizeResponse.assert_called_once_with("Hello", {"tone": "friendly"})

    @patch('src.main.translateMessage')
    def test_translateMessage(self, mock_translateMessage):
        main.translateMessage("Hello", "fr")
        mock_translateMessage.assert_called_once_with("Hello", "fr")

    @patch('src.main.maintainContext')
    def test_maintainContext(self, mock_maintainContext):
        main.maintainContext("Hello")
        mock_maintainContext.assert_called_once_with("Hello")

    @patch('src.main.learnAndImprove')
    def test_learnAndImprove(self, mock_learnAndImprove):
        main.learnAndImprove("Hello", "Bonjour")
        mock_learnAndImprove.assert_called_once_with("Hello", "Bonjour")

    @patch('src.main.ensurePrivacy')
    def test_ensurePrivacy(self, mock_ensurePrivacy):
        main.ensurePrivacy("Hello")
        mock_ensurePrivacy.assert_called_once_with("Hello")

    @patch('src.main.renderUI')
    def test_renderUI(self, mock_renderUI):
        main.renderUI()
        mock_renderUI.assert_called_once()

    @patch('src.main.handleAPIRequest')
    def test_handleAPIRequest(self, mock_handleAPIRequest):
        main.handleAPIRequest("GET", "/api/chat")
        mock_handleAPIRequest.assert_called_once_with("GET", "/api/chat")

    @patch('src.main.storeData')
    def test_storeData(self, mock_storeData):
        main.storeData({"message": "Hello"})
        mock_storeData.assert_called_once_with({"message": "Hello"})

    @patch('src.main.deployApplication')
    def test_deployApplication(self, mock_deployApplication):
        main.deployApplication()
        mock_deployApplication.assert_called_once()

    @patch('src.main.sendNotification')
    def test_sendNotification(self, mock_sendNotification):
        main.sendNotification("New message")
        mock_sendNotification.assert_called_once_with("New message")

    @patch('src.main.authenticateUser')
    def test_authenticateUser(self, mock_authenticateUser):
        main.authenticateUser("username", "password")
        mock_authenticateUser.assert_called_once_with("username", "password")

    @patch('src.main.runCICD')
    def test_runCICD(self, mock_runCICD):
        main.runCICD()
        mock_runCICD.assert_called_once()

if __name__ == '__main__':
    unittest.main()
```